Fix compile error

+++ mongoose.c
@@ -3051,7 +3051,7 @@ MG_INTERNAL struct mg_connection *mg_do_connect(struct mg_connection *nc,
   LOG(LL_DEBUG, ("%p %s://%s:%hu", nc, proto == SOCK_DGRAM ? "udp" : "tcp",
                  inet_ntoa(sa->sin.sin_addr), ntohs(sa->sin.sin_port)));
 #ifdef __LINUX_SOCKETCAN__
-# error TODO:
+# warning TODO: implement mg_do_connect() for SocketCAN
 #endif
   nc->flags |= MG_F_CONNECTING;
   if (proto == SOCK_DGRAM) {
@@ -3902,7 +3902,7 @@ static int mg_socket_if_can_send(struct mg_connection *nc, const void *buf,
   int mtu = sizeof(frame);
   size_t max_dlen = CAN_MAX_DLEN;
   fd_set fds;
-  int can_id = nc->mgr->user_data;
+  int can_id = (int)nc->mgr->user_data;
   int i;
   int rc = 0;
